backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1

# File Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# 3D Processing & Rendering
trimesh==4.0.5
pyrender==0.1.45
Open3D==0.18.0

# Plan2Scene Dependencies
torch==2.1.1
torchvision==0.16.1
scipy==1.11.4
scikit-image==0.22.0
matplotlib==3.8.2
pandas==2.1.3

# API & Storage
boto3==1.29.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
